+++ raster_tools/extract.py
@@ -1,6 +1,12 @@
 # -*- coding: utf-8 -*-
 # (c) Nelen & Schuurmans, see LICENSE.rst.
-""" Extract layers from a raster server using a geometry. """
+"""
+Extract layers from a raster server using a geometry.
+
+Main operation is to build and save a rudimentary 2D part for a 3Di model.
+Alternatively it is possible to extract from a single layer on the remote
+server.
+"""
 
 from __future__ import print_function
 from __future__ import unicode_literals
@@ -10,9 +16,7 @@ from __future__ import division
 import argparse
 import collections
 import csv
-import logging
 import os
-import sys
 import threading
 
 try:
@@ -34,14 +38,10 @@ from osgeo import gdalnumeric as np
 from osgeo import ogr
 from osgeo import osr
 
-logger = logging.getLogger(__name__)
-gdal.UseExceptions()
-ogr.UseExceptions()
-osr.UseExceptions()
 operations = {}
 
 # Version management for outdated warning
-VERSION = 20
+VERSION = 21
 
 GITHUB_URL = ('https://raw.github.com/nens/'
               'raster-tools/master/raster_tools/extract.py')
@@ -64,7 +64,7 @@ ATTRIBUTE = 'model'
 CELLSIZE = 0.5, 0.5
 DTYPE = 'f4'
 FLOOR = 0.15
-OPERATION = '3di'
+OPERATION = '3di-ahn2'
 PROJECTION = 'EPSG:28992'
 TIMESTAMP = '1970-01-01T00:00:00Z'
 SERVER = 'https://raster.lizard.net'
@@ -126,10 +126,8 @@ class Layers(Operation):
         return {self.name: result}
 
 
-class ThreeDi(Operation):
+class ThreeDiBase(object):
     """ Extract all rasters for a 3Di model. """
-    name = '3di'
-
     # inputs
     I_SOIL = 'soil'
     I_LANDUSE = 'landuse'
@@ -177,17 +175,10 @@ class ThreeDi(Operation):
     required = set([y for x in outputs.values() for y in x])
 
     def __init__(self, floor, landuse, soil, time, **kwargs):
-        """ Initialize the operation. """
-        self.layers = {
-            self.I_BATHYMETRY: dict(layers=','.join([
-                'ahn2:int',
-                'dem:bag!{}'.format(floor),
-                'dem:water',
-            ])),
-            self.I_LANDUSE: dict(layers='cover:3di'),
-            self.I_SOIL: dict(layers='soil:3di'),
-        }
-
+        """
+        Initialize the operation. The subclasses have to set the layers
+        attribute.
+        """
         # create the inputs that will be loaded from the server
         self.inputs = {}
         for k in self.layers:
@@ -479,6 +470,44 @@ class ThreeDi(Operation):
         return {key: self.calculators[key](datasets) for key in self.outputs}
 
 
+class ThreeDiAHN2(Operation, ThreeDiBase):
+    name = '3di-ahn2'
+
+    def __init__(self, floor, **kwargs):
+        """ Initialize the operation. """
+        # first define the layers
+        self.layers = {
+            self.I_BATHYMETRY: dict(layers=','.join([
+                'intern:nl:ahn2:int',
+                'intern:nl:ahn2:bag!{}'.format(floor),
+                'dem:water',
+            ])),
+            self.I_LANDUSE: dict(layers='cover:3di'),
+            self.I_SOIL: dict(layers='soil:3di'),
+        }
+        # then let the base class handle the rest
+        super(ThreeDiAHN2, self).__init__(floor, **kwargs)
+
+
+class ThreeDiAHN3(Operation, ThreeDiBase):
+    name = '3di-ahn3'
+
+    def __init__(self, floor, **kwargs):
+        """ Initialize the operation. """
+        # first define the layers
+        self.layers = {
+            self.I_BATHYMETRY: dict(layers=','.join([
+                'intern:nl:ahn3:int',
+                'intern:nl:ahn3:bag!{}'.format(floor),
+                'dem:water',
+            ])),
+            self.I_LANDUSE: dict(layers='cover:3di'),
+            self.I_SOIL: dict(layers='soil:3di'),
+        }
+        # then let the base class handle the rest
+        super(ThreeDiAHN3, self).__init__(floor, **kwargs)
+
+
 class Preparation(object):
     """
     Preparation.
@@ -538,7 +567,7 @@ class Preparation(object):
         return geometry
 
     def _create_dataset(self, name, path):
-        """ """
+        """ Prepare output tif dataset. """
         # dir
         try:
             os.makedirs(os.path.dirname(path))
@@ -565,6 +594,11 @@ class Preparation(object):
         dataset.GetRasterBand(1).SetNoDataValue(
             self.operation.no_data_value[name],
         )
+
+        # meta
+        meta = {n.upper(): self.operation.inputs[n]['layers']
+                for n in self.operation.outputs[name]}
+        dataset.SetMetadata(meta)
         return dataset
 
     def _get_or_create_datasets(self):
@@ -1011,9 +1045,7 @@ def get_parser():
 
 def main():
     """ Call command with args from parser. """
-    operations.update({cls.name: cls
-                       for cls in Operation.__subclasses__()})
-    logging.basicConfig(stream=sys.stderr, level=logging.DEBUG)
+    operations.update({cls.name: cls for cls in Operation.__subclasses__()})
     return command(**vars(get_parser().parse_args()))